refactor: split better rp module
