+++ modules/mod-better-roll-play/CMakeLists.txt
@@ -1,5 +1,10 @@
 AC_ADD_SCRIPT_MODULE(mod_better_roll_play
     SOURCES
-        mod_better_roll_play.cpp
+        src/BetterRPConfig.cpp
+        src/BetterRPState.cpp
+        src/BetterRPModule.cpp
+        src/BetterRPCommands.cpp
+    PRIVATE_INCLUDES
+        src
 )
